Perubahan pertama

+++ src/pages/api/categories/[id].js
@@ -2,6 +2,7 @@
 import db from "@/lib/mongoDB";
 import auth from "@/middleware/auth";
 import Category from "@/models/Category";
+import Product from "@/models/Product";
 
 db.connect();
 
@@ -46,6 +47,12 @@ const deleteCategory = async (req, res) => {
       return res.status(400).json({ err: "Authentication is not valid." });
 
     const { id } = req.query;
+
+    const products = await Product.findOne({ category: id });
+    if (products)
+      return res.status(400).json({
+        err: "Hapus semua produk yang memiliki hubungan dengan kategori ini",
+      });
 
     await Category.findByIdAndDelete(id);
     res.json({ msg: "Hapus kategori berhasil!" });
